reactions/dapr-agents-router/pyproject.toml
[project]
name = "drasi-dapr-agents-router"
version = "0.1.0"
description = "Drasi Reaction that routes change events to Dapr Agents via Pub/Sub"
requires-python = ">=3.11"
dependencies = [
    "drasi-reaction-sdk>=0.1.0",
    "dapr>=1.14.0",
    "pydantic>=2.10.0",
    "pyyaml>=6.0",
    "mcp>=1.0.0",
    "uvicorn>=0.27.0",
    "fastapi>=0.110.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
